# Conflicts:
#	Backend/src/main/java/com/unesco/core/controller/DepartmentController.java
#	Backend/src/main/java/com/unesco/core/controller/GroupController.java
#	Backend/src/main/java/com/unesco/core/controller/PairController.java
#	Backend/src/main/java/com/unesco/core/controller/ProfessorController.java
#	Backend/src/main/java/com/unesco/core/entities/Department.java
#	Backend/src/main/java/com/unesco/core/entities/schedule/Pair.java
#	Backend/src/main/java/com/unesco/core/entities/schedule/Room.java
#	Backend/src/main/java/com/unesco/core/models/PairModel.java

+++ Backend/src/main/java/com/unesco/core/entities/schedule/Room.java
@@ -18,7 +18,7 @@ public class Room implements LongId {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
